GCP/nginx/requirements.txt
Django
numpy
gunicorn==19.6.0
google-cloud-datastore==1.7.0